+++ pymsis/utils.py
@@ -1,5 +1,6 @@
 """Utilities for obtaining input datasets."""
 
+import os
 import urllib.request
 import warnings
 from io import BytesIO
@@ -13,8 +14,45 @@ import pymsis
 
 _DATA_FNAME: str = "SW-All.csv"
 _F107_AP_URL: str = f"https://celestrak.org/SpaceData/{_DATA_FNAME}"
-_F107_AP_PATH: Path = Path(pymsis.__file__).parent / _DATA_FNAME
+_F107_AP_DEFAULT_PATH: Path = Path(pymsis.__file__).parent / _DATA_FNAME
 _DATA: dict[str, npt.NDArray] | None = None
+
+_F107_AP_PATH: Path = Path(
+    os.environ.get("PYMSIS_SPACE_WEATHER_PATH", _F107_AP_DEFAULT_PATH)
+)
+
+
+def set_space_weather_path(path: str | Path) -> None:
+    """
+    Set a custom path to retrieve the F10.7 and ap data from.
+
+    By default, the data is retrieved from CelesTrak and stored inside the
+    installed package location. Retrieving the data from a custom path may
+    be useful if you are retrieving the data from a different source, if you
+    have a centralized location for the data, or would like to use custom data.
+
+    Alternatively, you can set the environment variable ``PYMSIS_SPACE_WEATHER_PATH``
+    to the desired path to achieve the same result and to avoid setting the space
+    weather path programmatically on each run.
+
+    Setting a default and running `download_f107_ap()` will not download to
+    this custom path.
+
+    Parameters
+    ----------
+    path : str or Path
+        Path to the F10.7 and ap data file, retrieved from CelesTrak or elsewhere.
+    """
+    if not Path(path).exists():
+        raise FileNotFoundError(
+            f"Provided custom space weather path does not exist: {path}"
+        )
+
+    # set the global variable to the new path
+    globals()["_F107_AP_PATH"] = Path(path)
+
+    # reset the global data
+    globals()["_DATA"] = None
 
 
 def download_f107_ap() -> None:
@@ -25,6 +63,9 @@ def download_f107_ap() -> None:
     the same filename as the data source: ``SW-All.csv``.
     This routine can be called to update the data as well if you would like to
     use newer data since the last time you downloaded the file.
+
+    If `set_space_weather_path()` has been called to set a custom path, the file will
+    still be downloaded to the default location and thus ignored by pymsis.
 
     Notes
     -----
@@ -44,13 +85,21 @@ def download_f107_ap() -> None:
     """
     warnings.warn(f"Downloading ap and F10.7 data from {_F107_AP_URL}")
     req = urllib.request.urlopen(_F107_AP_URL)
-    with _F107_AP_PATH.open("wb") as f:
+    with _F107_AP_DEFAULT_PATH.open("wb") as f:
         f.write(req.read())
 
 
 def _load_f107_ap_data() -> dict[str, npt.NDArray]:
     """Load data from disk, if it isn't present go out and download it first."""
-    if not _F107_AP_PATH.exists():
+    default_file_exists = _F107_AP_DEFAULT_PATH.exists()
+    custom_file_used = _F107_AP_PATH != _F107_AP_DEFAULT_PATH
+
+    if custom_file_used and not _F107_AP_PATH.exists():
+        raise FileNotFoundError(
+            f"Custom space weather path has been set but does not exist: {_F107_AP_PATH}"
+        )
+
+    if not custom_file_used and not default_file_exists:
         download_f107_ap()
 
     dtype = {